// src/app/models/medico_response.model.ts
// Generated by https://quicktype.io

export interface MedicoResponse {
    exito:   boolean;
    medicos: Medico[];
}

export interface Medico {
    _id:      string;
    nombre:   string;
    usuario:  MedicoObjeto;
    hospital: MedicoObjeto;
    img:      string;
}

interface MedicoObjeto {
    _id:    string;
    nombre: string;
    img:    string;
}